Use consistent topology name - EndpointOrientedMigrationTopology
